reconstruct the update framework
