+++ tests/test_update.py
@@ -1,7 +1,90 @@
 import tempfile
 from pathlib import Path
 
-from chemsmart.cli.update import Updater
+from click.testing import CliRunner
+
+from chemsmart.cli.update import (
+    DependencyUpdater,
+    VersionUpdater,
+    update,
+)
+
+
+def test_dependency_cli_delegates_to_dependency_updater(monkeypatch):
+    called = False
+
+    class FakeDependencyUpdater:
+        def update(self):
+            nonlocal called
+            called = True
+
+    monkeypatch.setattr(
+        "chemsmart.cli.update.DependencyUpdater", FakeDependencyUpdater
+    )
+
+    result = CliRunner().invoke(update, ["deps"])
+
+    assert result.exit_code == 0, result.output
+    assert called
+
+
+def test_version_cli_constructs_version_updater_with_cli_value(monkeypatch):
+    seen_version = None
+    called = False
+
+    class FakeVersionUpdater:
+        def __init__(self, version_number):
+            nonlocal seen_version
+            seen_version = version_number
+
+        def update(self):
+            nonlocal called
+            called = True
+
+    monkeypatch.setattr(
+        "chemsmart.cli.update.VersionUpdater", FakeVersionUpdater
+    )
+
+    result = CliRunner().invoke(
+        update, ["version", "--version-number", "9.8.7"]
+    )
+
+    assert result.exit_code == 0, result.output
+    assert seen_version == "9.8.7"
+    assert called
+
+
+def test_version_updater_updates_version_files_in_order(tmp_path):
+    version_file = tmp_path / "VERSION"
+    pyproject_file = tmp_path / "pyproject.toml"
+    docs_conf_file = tmp_path / "docs" / "source" / "conf.py"
+    docs_conf_file.parent.mkdir(parents=True)
+
+    version_file.write_text("0.1.9\n", encoding="utf-8")
+    pyproject_file.write_text(
+        'name = "chemsmart"\nversion = "0.1.9"  # keep pyproject comment\n',
+        encoding="utf-8",
+    )
+    docs_conf_file.write_text(
+        'project = "CHEMSMART"\nrelease = "0.1.9"  # keep docs comment\n',
+        encoding="utf-8",
+    )
+
+    updater = VersionUpdater("4.5.6")
+    updater._version_file_path = version_file
+    updater._pyproject_path = pyproject_file
+    updater._docs_conf_file_path = docs_conf_file
+
+    updater.update()
+
+    assert version_file.read_text(encoding="utf-8") == "4.5.6\n"
+    assert (
+        pyproject_file.read_text(encoding="utf-8")
+        == 'name = "chemsmart"\nversion = "4.5.6"  # keep pyproject comment\n'
+    )
+    assert docs_conf_file.read_text(encoding="utf-8") == (
+        'project = "CHEMSMART"\nrelease = "4.5.6"  # keep docs comment\n'
+    )
 
 
 class TestUpdater:
@@ -9,7 +92,7 @@ class TestUpdater:
 
     def test_extract_pkg_name_normalization(self):
         """Test that package names are normalized correctly (PEP 503)."""
-        updater = Updater()
+        updater = DependencyUpdater()
 
         # Create a temporary requirements file
         with tempfile.NamedTemporaryFile(
@@ -53,7 +136,7 @@ class TestUpdater:
     def test_underscore_dash_equivalence(self):
         """Test that underscore and dash package
         names are treated as equivalent."""
-        updater = Updater()
+        updater = DependencyUpdater()
 
         # Create a temporary requirements file with underscore
         with tempfile.NamedTemporaryFile(
@@ -85,7 +168,7 @@ class TestUpdater:
     def test_dash_underscore_equivalence(self):
         """Test that dash and underscore package
         names are treated as equivalent."""
-        updater = Updater()
+        updater = DependencyUpdater()
 
         # Create a temporary requirements file with dash
         with tempfile.NamedTemporaryFile(
@@ -116,7 +199,7 @@ class TestUpdater:
 
     def test_new_dependency_detected(self):
         """Test that genuinely new dependencies are still detected."""
-        updater = Updater()
+        updater = DependencyUpdater()
 
         # Create a temporary requirements file
         with tempfile.NamedTemporaryFile(
